runpod_service/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# WebSocket
websockets>=12.0

# Image processing
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0

# AI/ML - GPU
onnxruntime-gpu>=1.17.0
insightface>=0.7.3

# FaceFusion core dependencies
onnx>=1.15.0
protobuf>=4.25.0

# Face enhancement
gfpgan>=1.3.8
basicsr>=1.4.2

# Utilities
tqdm>=4.66.0
requests>=2.31.0
aiofiles>=23.2.1
